fix(tests): split and accelerate post-build-review tests
- Split 19 genuine review pipeline tests into tests/audit/test_post_build_review_pipeline.py
- Replaced redundant packet preparations with module-scoped fixtures (bilash_packet, malyshko_packet, malyshko_exact_only_packet, a1_packet)
- Added minimal synthetic packet fixture for sub-millisecond contract and schema testing
- Reduced runtime from 529s on a single test file to ~110s and ~119s across two parallelizable files

Fixes #7001

Entire-Checkpoint: 310bb6eb4f82
