+++ tests/fixtures/post_build_review/synthetic/activities.md
@@ -1,0 +1,2 @@
+Вправа 1: синтетичний аналіз тексту та структури.
+Вправа 2: додатковий приклад для практики.
