+++ tests/fixtures/post_build_review/synthetic/content.md
@@ -1,0 +1,5 @@
+Synthetic quality-dimension evidence for the review contract.
+A second independent anchor supports exceptional-score validation.
+Кожен приклад має спиратися на перевірені джерела інформації.
+Розрізніть, що встановлює каталог Держкіно, а що міг би додати сам фільм.
+:::
