+++ tests/fixtures/post_build_review/synthetic/plan.yaml
@@ -1,0 +1,1 @@
+plan: synthetic-fixture
