+++ tests/fixtures/post_build_review/synthetic/resources.yaml
@@ -1,0 +1,2 @@
+- title: Resource 1
+  url: https://example.com/res1
