+++ tests/fixtures/post_build_review/synthetic/vocabulary.yaml
@@ -1,0 +1,3 @@
+- lemma: синтетичний
+- lemma: додатковий
+- lemma: відсутній
